+++ 1.20/Forge/src/main/java/fuzs/hangglider/HangGliderForge.java
@@ -1,25 +1,19 @@
 package fuzs.hangglider;
 
 import fuzs.hangglider.capability.GlidingCapability;
-import fuzs.hangglider.data.ModModelProvider;
 import fuzs.hangglider.data.ModLanguageProvider;
+import fuzs.hangglider.data.ModModelProvider;
 import fuzs.hangglider.data.ModRecipeProvider;
 import fuzs.hangglider.data.ModSpriteSourceProvider;
 import fuzs.hangglider.init.ModRegistry;
 import fuzs.puzzleslib.api.capability.v2.ForgeCapabilityHelper;
 import fuzs.puzzleslib.api.core.v1.ContentRegistrationFlags;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
-import net.minecraft.core.HolderLookup;
-import net.minecraft.data.DataGenerator;
-import net.minecraft.data.PackOutput;
 import net.minecraftforge.common.capabilities.CapabilityToken;
-import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.data.event.GatherDataEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
-
-import java.util.concurrent.CompletableFuture;
 
 @Mod(HangGlider.MOD_ID)
 @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD)
@@ -27,7 +21,7 @@ public class HangGliderForge {
 
     @SubscribeEvent
     public static void onConstructMod(final FMLConstructModEvent evt) {
-        ModConstructor.construct(HangGlider.MOD_ID, HangGlider::new, ContentRegistrationFlags.LEGACY_SMITHING);
+        ModConstructor.construct(HangGlider.MOD_ID, HangGlider::new, ContentRegistrationFlags.COPY_TAG_RECIPES);
         registerCapabilities();
     }
 
@@ -37,13 +31,9 @@ public class HangGliderForge {
 
     @SubscribeEvent
     public static void onGatherData(final GatherDataEvent evt) {
-        final DataGenerator dataGenerator = evt.getGenerator();
-        final PackOutput packOutput = dataGenerator.getPackOutput();
-        final CompletableFuture<HolderLookup.Provider> lookupProvider = evt.getLookupProvider();
-        final ExistingFileHelper fileHelper = evt.getExistingFileHelper();
-        dataGenerator.addProvider(true, new ModModelProvider(packOutput, HangGlider.MOD_ID, fileHelper));
-        dataGenerator.addProvider(true, new ModLanguageProvider(packOutput, HangGlider.MOD_ID));
-        dataGenerator.addProvider(true, new ModRecipeProvider(packOutput));
-        dataGenerator.addProvider(true, new ModSpriteSourceProvider(packOutput, fileHelper));
+        evt.getGenerator().addProvider(true, new ModModelProvider(evt, HangGlider.MOD_ID));
+        evt.getGenerator().addProvider(true, new ModLanguageProvider(evt, HangGlider.MOD_ID));
+        evt.getGenerator().addProvider(true, new ModRecipeProvider(evt, HangGlider.MOD_ID));
+        evt.getGenerator().addProvider(true, new ModSpriteSourceProvider(evt, HangGlider.MOD_ID));
     }
 }