+++ 1.20/Forge/src/main/java/fuzs/hangglider/data/ModLanguageProvider.java
@@ -3,12 +3,12 @@ package fuzs.hangglider.data;
 import fuzs.hangglider.HangGlider;
 import fuzs.hangglider.init.ModRegistry;
 import fuzs.puzzleslib.api.data.v1.AbstractLanguageProvider;
-import net.minecraft.data.PackOutput;
+import net.minecraftforge.data.event.GatherDataEvent;
 
 public class ModLanguageProvider extends AbstractLanguageProvider {
 
-    public ModLanguageProvider(PackOutput packOutput, String modId) {
-        super(packOutput, modId);
+    public ModLanguageProvider(GatherDataEvent evt, String modId) {
+        super(evt, modId);
     }
 
     @Override