+++ 1.20/Forge/src/main/java/fuzs/hangglider/data/ModModelProvider.java
@@ -2,14 +2,13 @@ package fuzs.hangglider.data;
 
 import fuzs.hangglider.init.ModRegistry;
 import fuzs.puzzleslib.api.data.v1.AbstractModelProvider;
-import net.minecraft.data.PackOutput;
 import net.minecraftforge.client.model.generators.ModelFile;
-import net.minecraftforge.common.data.ExistingFileHelper;
+import net.minecraftforge.data.event.GatherDataEvent;
 
 public class ModModelProvider extends AbstractModelProvider {
 
-    public ModModelProvider(PackOutput packOutput, String modId, ExistingFileHelper fileHelper) {
-        super(packOutput, modId, fileHelper);
+    public ModModelProvider(GatherDataEvent evt, String modId) {
+        super(evt, modId);
     }
 
     @Override