+++ 1.20/Forge/src/main/java/fuzs/hangglider/data/ModRecipeProvider.java
@@ -1,31 +1,31 @@
 package fuzs.hangglider.data;
 
-import fuzs.hangglider.HangGlider;
 import fuzs.hangglider.init.ModRegistry;
 import fuzs.puzzleslib.api.data.v1.AbstractRecipeProvider;
-import net.minecraft.data.PackOutput;
+import fuzs.puzzleslib.api.data.v1.recipes.CopyTagShapelessRecipeBuilder;
 import net.minecraft.data.recipes.FinishedRecipe;
 import net.minecraft.data.recipes.RecipeCategory;
 import net.minecraft.data.recipes.ShapedRecipeBuilder;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraftforge.data.event.GatherDataEvent;
 
 import java.util.function.Consumer;
 
 public class ModRecipeProvider extends AbstractRecipeProvider {
 
-    public ModRecipeProvider(PackOutput packOutput) {
-        super(packOutput);
+    public ModRecipeProvider(GatherDataEvent evt, String modId) {
+        super(evt, modId);
     }
 
     @Override
     protected void buildRecipes(Consumer<FinishedRecipe> recipeConsumer) {
         ShapedRecipeBuilder.shaped(RecipeCategory.TOOLS, ModRegistry.GLIDER_WING_ITEM.get())
                 .define('S', Items.STICK)
-                .define('#', Items.LEATHER)
-                .define('P', Items.PHANTOM_MEMBRANE)
-                .pattern(" SP")
-                .pattern("SP#")
-                .pattern("P##")
+                .define('#', Ingredient.of(Items.LEATHER, Items.RABBIT_HIDE))
+                .pattern("  S")
+                .pattern(" S#")
+                .pattern("S##")
                 .unlockedBy(getHasName(Items.LEATHER), has(Items.LEATHER))
                 .save(recipeConsumer);
         ShapedRecipeBuilder.shaped(RecipeCategory.TOOLS, ModRegistry.GLIDER_FRAMEWORK_ITEM.get())
@@ -41,6 +41,11 @@ public class ModRecipeProvider extends AbstractRecipeProvider {
                 .pattern("#@#")
                 .unlockedBy(getHasName(ModRegistry.GLIDER_FRAMEWORK_ITEM.get()), has(ModRegistry.GLIDER_FRAMEWORK_ITEM.get()))
                 .save(recipeConsumer);
-        legacyNetheriteSmithing(HangGlider.MOD_ID, recipeConsumer, ModRegistry.HANG_GLIDER_ITEM.get(), Items.ELYTRA, RecipeCategory.TOOLS, ModRegistry.REINFORCED_HANG_GLIDER_ITEM.get());
+        CopyTagShapelessRecipeBuilder.shapeless(RecipeCategory.TOOLS, ModRegistry.REINFORCED_HANG_GLIDER_ITEM.get())
+                .requires(ModRegistry.HANG_GLIDER_ITEM.get())
+                .requires(Items.ELYTRA)
+                .copyFrom(ModRegistry.HANG_GLIDER_ITEM.get())
+                .unlockedBy(getHasName(Items.ELYTRA), has(Items.ELYTRA))
+                .save(recipeConsumer);
     }
 }