+++ 1.20/Forge/src/main/java/fuzs/hangglider/data/ModSpriteSourceProvider.java
@@ -3,16 +3,15 @@ package fuzs.hangglider.data;
 import fuzs.hangglider.client.handler.ElytraEquippedHandler;
 import fuzs.puzzleslib.api.data.v1.AbstractSpriteSourceProvider;
 import net.minecraft.client.renderer.texture.atlas.sources.SingleFile;
-import net.minecraft.data.PackOutput;
-import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.common.data.SpriteSourceProvider;
+import net.minecraftforge.data.event.GatherDataEvent;
 
 import java.util.Optional;
 
 public class ModSpriteSourceProvider extends AbstractSpriteSourceProvider {
 
-    public ModSpriteSourceProvider(PackOutput packOutput, ExistingFileHelper fileHelper) {
-        super(packOutput, fileHelper);
+    public ModSpriteSourceProvider(GatherDataEvent evt, String modId) {
+        super(evt, modId);
     }
 
     @Override